requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
blinker==1.7.0
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cohere==4.56
colorama==0.4.6
cryptography==42.0.5
dashscope==1.20.14
dataclasses-json==0.6.4
distro==1.9.0
exceptiongroup==1.2.0
faiss-cpu==1.8.0
fastavro==1.9.4
filelock==3.16.1
filetype==1.2.0
frozenlist==1.4.1
fsspec==2024.12.0
gitdb==4.0.11
GitPython==3.1.42
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.158.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.27.1
idna==3.7
importlib-metadata==6.11.0
Jinja2==3.1.3
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-google-genai==2.0.8
langchain-ollama==0.2.2
langchain-openai==0.3.0
langchain-text-splitters==0.3.5
langsmith==0.2.10
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.4.2
numpy==1.26.4
ollama==0.4.5
openai==1.59.6
orjson==3.10.14
packaging==24.2
pandas==2.2.3
pdfminer.six==20231228
pillow==11.1.0
propcache==0.2.1
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==15.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.32.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
setuptools==75.6.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.28
streamlit==1.41.1
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
toolz==0.12.1
torch==2.5.1
tornado==6.4.2
tqdm==4.66.3
transformers==4.48.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
validators==0.22.0
watchdog==4.0.0
websocket-client==1.8.0
wheel==0.45.1
yarl==1.18.3
zhipuai==2.1.5.20250106
zipp==3.19.1
